requirements.txt
numpy 1.18.1
imageio 2.8.0
matplotlib 3.1.3
tensorflow-gpu 1.12.0
sklearn 0.0
scikit-image 0.16.2
scikit-learn 0.22.2.post1
scipy 1.1.0
Pillow 6.2.0
pandas 1.0.1
opencv-python 3.4.4.19
Keras 2.2.4
tabulate 0.8.7